requirements.txt
torch==1.7.1
torchvision==0.8.2
numpy==1.19.2
Pillow==8.1.0
